refactor: MediaTypes 분리 및 불필요한 클래스 삭제
